+++ src/fileFrom/types.ts
@@ -1,20 +1,43 @@
-import {UploadcareFileInterface} from '../types'
-import {UploadInterface} from '../lifecycle/types'
+import {FileData} from '../types'
+import {Url, Uuid} from '..'
+import {isNode} from '../tools/isNode'
 
-export enum FileFromEnum {
-  Object = 'object',
-  URL = 'url',
-  DOM = 'input',
-  Uploaded = 'uploaded',
+/**
+ * FileData type guard.
+ *
+ * @param {FileData | Url | Uuid} data
+ */
+export const isFileData = (data: FileData | Url | Uuid): data is FileData => {
+  return data !== undefined &&
+    (
+      (!isNode() && data instanceof Blob) ||
+      (!isNode() && data instanceof File) ||
+      (isNode() && data instanceof Buffer)
+    )
 }
 
 /**
- * Base `thenable` interface for uploading `fileFrom` (`object`, `url`, `input`, `uploaded`).
+ * Uuid type guard.
+ *
+ * @param {FileData | Url | Uuid} data
  */
-export interface FileUploadInterface extends UploadInterface<UploadcareFileInterface> {}
+export const isUuid = (data: FileData | Url | Uuid): data is Uuid => {
+  const UUID_REGEX = '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'
+  const regExp = (new RegExp(UUID_REGEX))
 
-export interface HandlerInterface<T> {
-  upload(): Promise<T>
+  return !isFileData(data) &&
+    regExp.test(data)
 }
 
-export interface FileHandlerInterface extends HandlerInterface<UploadcareFileInterface> {}
+/**
+ * Url type guard.
+ *
+ * @param {FileData | Url | Uuid} data
+ */
+export const isUrl = (data: FileData | Url | Uuid): data is Url => {
+  const URL_REGEX = '^(?:\\w+:)?\\/\\/([^\\s\\.]+\\.\\S{2}|localhost[\\:?\\d]*)\\S*$'
+  const regExp = (new RegExp(URL_REGEX))
+
+  return !isFileData(data) &&
+    regExp.test(data)
+}