+++ src/groupFrom/groupFrom.ts
@@ -1,31 +1,51 @@
-import {UploadFromObject} from './UploadFromObject'
-import {UploadFromUrl} from './UploadFromUrl'
-import {UploadFromUploaded} from './UploadFromUploaded'
+import {GroupUploadLifecycle} from '../lifecycle/GroupUploadLifecycle'
+import {UploadLifecycle} from '../lifecycle/UploadLifecycle'
+import {GroupFromObject} from './GroupFromObject'
+import {GroupFromUploaded} from './GroupFromUploaded'
+import {GroupFromUrl} from './GroupFromUrl'
 
 /* Types */
-import {FileData, SettingsInterface} from '../types'
+import {FileData, SettingsInterface, UploadcareGroupInterface} from '../types'
 import {Url} from '../api/fromUrl'
 import {Uuid} from '../api/types'
-import {GroupFromEnum, GroupUploadInterface} from './types'
+import {isFileDataArray, isUrlArray, isUuidArray} from './types'
+import {GroupUploadLifecycleInterface, UploadInterface} from '../lifecycle/types'
+import {Upload} from '../lifecycle/Upload'
+import {createProxyHandler} from '../lifecycle/createProxyHandler'
 
 /**
  * Uploads file from provided data.
  *
- * @param {FileFromEnum} from
  * @param {FileData} data
  * @param {SettingsInterface} settings
  * @throws Error
- * @returns {FileUploadInterface}
+ * @returns {UploadInterface<UploadcareGroupInterface>}
  */
-export default function groupFrom(from: GroupFromEnum, data: FileData[] | Url[] | Uuid[], settings: SettingsInterface = {}): GroupUploadInterface {
-  switch (from) {
-    case GroupFromEnum.Object:
-      return new UploadFromObject(data as FileData[], settings)
-    case GroupFromEnum.URL:
-      return new UploadFromUrl(data as Url[], settings)
-    case GroupFromEnum.Uploaded:
-      return new UploadFromUploaded(data as Uuid[], settings)
-    default:
-      throw new TypeError(`Group uploading from "${from}" is not supported`)
+export default function groupFrom(data: FileData[] | Url[] | Uuid[], settings: SettingsInterface = {}): UploadInterface<UploadcareGroupInterface> {
+  const lifecycle = new UploadLifecycle<UploadcareGroupInterface>()
+  const groupUploadLifecycle = new GroupUploadLifecycle(lifecycle)
+  const lifecycleProxyHandler = createProxyHandler<UploadcareGroupInterface>(lifecycle)
+
+  if (isFileDataArray(data)) {
+    const fileHandler = new GroupFromObject(data, settings)
+    const fileUpload = new Upload<UploadcareGroupInterface, GroupUploadLifecycleInterface>(groupUploadLifecycle, fileHandler)
+
+    return new Proxy(fileUpload, lifecycleProxyHandler)
   }
+
+  if (isUrlArray(data)) {
+    const fileHandler = new GroupFromUrl(data, settings)
+    const fileUpload = new Upload<UploadcareGroupInterface, GroupUploadLifecycleInterface>(groupUploadLifecycle, fileHandler)
+
+    return new Proxy(fileUpload, lifecycleProxyHandler)
+  }
+
+  if (isUuidArray(data)) {
+    const fileHandler = new GroupFromUploaded(data, settings)
+    const fileUpload = new Upload<UploadcareGroupInterface, GroupUploadLifecycleInterface>(groupUploadLifecycle, fileHandler)
+
+    return new Proxy(fileUpload, lifecycleProxyHandler)
+  }
+
+  throw new TypeError(`Group uploading from "${data}" is not supported`)
 }