+++ src/lifecycle/FileUploadLifecycle.ts
@@ -1,39 +1,27 @@
+import checkFileIsReady from '../checkFileIsReady'
+import {UploadcareFile} from '../UploadcareFile'
+
+/* Types */
 import {FileUploadLifecycleInterface, LifecycleInterface} from './types'
 import {SettingsInterface, UploadcareFileInterface} from '../types'
 import {Uuid} from '..'
-import checkFileIsReady from '../checkFileIsReady'
-import prettyFileInfo from '../prettyFileInfo'
-import {UploadedState} from './state/UploadedState'
 import {PollPromiseInterface} from '../tools/poll'
 import {FileInfoInterface} from '../api/types'
 
 export class FileUploadLifecycle implements FileUploadLifecycleInterface {
-  private readonly lifecycle: LifecycleInterface<UploadcareFileInterface>
+  readonly uploadLifecycle: LifecycleInterface<UploadcareFileInterface>
   private isFileReadyPolling: PollPromiseInterface<FileInfoInterface> | null = null
 
   constructor(lifecycle: LifecycleInterface<UploadcareFileInterface>) {
-    this.lifecycle = lifecycle
+    this.uploadLifecycle = lifecycle
   }
 
   handleUploadedFile(uuid: Uuid, settings: SettingsInterface): Promise<UploadcareFileInterface> {
-    this.lifecycle.updateEntity({
-      uuid,
-      name: null,
-      size: null,
-      isStored: null,
-      isImage: null,
-      cdnUrl: null,
-      cdnUrlModifiers: null,
-      originalUrl: null,
-      originalFilename: null,
-      originalImageInfo: null,
-    })
+    const file = UploadcareFile.fromUuid(uuid)
+    const uploadLifecycle = this.uploadLifecycle
 
-    this.lifecycle.updateState(new UploadedState())
-
-    if (typeof this.lifecycle.onUploaded === 'function') {
-      this.lifecycle.onUploaded(uuid)
-    }
+    uploadLifecycle.updateEntity(file)
+    uploadLifecycle.handleUploaded(uuid)
 
     this.isFileReadyPolling = checkFileIsReady({
       uuid,
@@ -42,14 +30,14 @@ export class FileUploadLifecycle implements FileUploadLifecycleInterface {
 
     return this.isFileReadyPolling
       .then(info => {
-        this.lifecycle.updateEntity(prettyFileInfo(info, settings))
-
-        return Promise.resolve(this.lifecycle.getEntity())
+        const file = UploadcareFile.fromFileInfo(info, settings)
+        uploadLifecycle.updateEntity(file)
       })
-      .catch(error => Promise.reject(error))
+      .then(uploadLifecycle.handleReady.bind(uploadLifecycle))
+      .catch(uploadLifecycle.handleError.bind(uploadLifecycle))
   }
 
-  getUploadLifecycle(): LifecycleInterface<UploadcareFileInterface> {
-    return this.lifecycle
+  getIsFileReadyPolling = (): PollPromiseInterface<FileInfoInterface> | null => {
+    return this.isFileReadyPolling
   }
 }